Partial fix for window priority
Also fix a lot of warnings and add /W4

Ref: #14

+++ self_updater/unpacker.cpp
@@ -85,7 +85,7 @@ bool Unpacker::MemoryUnpack(const char* name, std::vector<Unpacker::FileContent>
 
 		file.name[nameLen] = '\0';
 
-		uint64_t fileSize = 0;
+		size_t fileSize = 0;
 		if (fread(&fileSize, sizeof(fileSize), 1, f) != 1) {
 			Logger::Error("Error while reading length of file\n");
 			return false;
